requirements.txt
django
requests==2.5.3
south
django-debug-toolbar
six
sqlparse
wsgiref
djangorestframework
django-rest-swagger
django-oauth-toolkit
markdown
django-filter
gunicorn
uwsgi
fabric
pyflakes
flake8
MySQL-python
coverage
Pillow
sparkpost
django-celery
easy-thumbnails
drf-nested-routers